fix param type

+++ src/main/java/com/example/demo/api/UserService.java
@@ -21,7 +21,7 @@ public interface UserService {
 
     long insertMultiUsers(List<User> list);
 
-    long deleteMultiUsersByUid(List<Integer> list);
+    long deleteMultiUsersByUid(List<Long> list);
 
     long updateMultiUser(List<User> list);
 }
